refactor: create hook for refining input
