+++ packages/frontend/src/pages/AiBuilder/components/AIFormModalContent.tsx
@@ -1,7 +1,6 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import { useEffect } from 'react'
 import { useForm } from 'react-hook-form'
 import { Form } from 'react-router-dom'
-import { useMutation } from '@apollo/client'
 import {
   Box,
   Button,
@@ -18,13 +17,13 @@ import {
 } from '@chakra-ui/react'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { FormLabel, useIsMobile } from '@opengovsg/design-system-react'
-import { debounce } from 'lodash'
 
 import pairLogo from '@/assets/pair-logo.svg'
 import { ImageBox } from '@/components/FlowStepConfigurationModal/ChooseAndAddConnection/ConfigureExcelConnection'
-import { REFINE_FORM_INPUT } from '@/graphql/mutations/ai/refine-form-input'
 import { AI_FORM_SCHEMA, AiFormData } from '@/pages/AiBuilder/schema'
 import { AI_FORM_IDEAS, AiFormIdea } from '@/pages/Flows/constants'
+
+import { useRefineFormInput } from '../hooks/useRefineFormInput'
 
 import IdeaButtons from './IdeaButtons'
 
@@ -65,9 +64,6 @@ export const AIFormModalContent = ({
   onBack: () => void
   onSubmit: (data: AiFormData) => void
 }) => {
-  const [refineFormInput, { loading: isRefiningFormInput }] =
-    useMutation(REFINE_FORM_INPUT)
-
   const isMobile = useIsMobile()
   const {
     register,
@@ -85,77 +81,26 @@ export const AIFormModalContent = ({
     },
   })
 
+  const triggerValue = watch('trigger')
   const actionsValue = watch('actions')
-  const isFromSuggestionRef = useRef(false)
-  const [aiSuggestion, setAiSuggestion] = useState<string | null>(null)
-  const [showReadyMessage, setShowReadyMessage] = useState(false)
 
-  const callRefineFormInput = useCallback(
-    async (prompt: string) => {
-      // Skip if the value is from a suggestion
-      if (isFromSuggestionRef.current) {
-        isFromSuggestionRef.current = false
-        return
-      }
-
-      // Skip if trigger is empty or too short (minimum 30 characters required by API)
-      if (!prompt || prompt.trim().length < 30) {
-        setAiSuggestion(null)
-        setShowReadyMessage(false)
-        return
-      }
-
-      try {
-        const result = await refineFormInput({
-          variables: {
-            input: {
-              prompt,
-              sessionId: null,
-            },
-          },
-        })
-
-        if (result.data?.refineFormInput) {
-          const { status, suggestion } = result.data.refineFormInput
-          if (status) {
-            // Status is true - show ready message
-            setShowReadyMessage(true)
-            setAiSuggestion(null)
-          } else if (suggestion) {
-            // Status is false and we have a suggestion
-            setAiSuggestion(suggestion)
-            setShowReadyMessage(false)
-          } else {
-            setAiSuggestion(null)
-            setShowReadyMessage(false)
-          }
-        }
-      } catch (error) {
-        console.error('Error refining form input:', error)
-        setAiSuggestion(null)
-        setShowReadyMessage(false)
-      }
-    },
-    [refineFormInput],
-  )
-
-  const debouncedRefineFormInput = useMemo(
-    () => debounce(callRefineFormInput, 2000),
-    [callRefineFormInput],
-  )
+  const {
+    refineFormInput,
+    isRefiningFormInput,
+    aiSuggestion,
+    showReadyMessage,
+    resetSuggestion,
+  } = useRefineFormInput()
 
   const handleIdeaClick = (idea: AiFormIdea) => {
-    isFromSuggestionRef.current = true
-    debouncedRefineFormInput.cancel()
-    setAiSuggestion(null)
-    setShowReadyMessage(false)
+    resetSuggestion()
     setValue('trigger', idea.trigger, { shouldValidate: true })
     setValue('actions', idea.actions, { shouldValidate: true })
   }
 
   useEffect(() => {
-    debouncedRefineFormInput(actionsValue)
-  }, [actionsValue, debouncedRefineFormInput])
+    refineFormInput(triggerValue, actionsValue)
+  }, [triggerValue, actionsValue, refineFormInput])
 
   const isCreate = type === 'create'
   const shouldShowIdeaButtons =
